+++ package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "@inrupt/lit-generated-vocab-common": "^0.3.11",
+    "axios": "^1.3.4"
+  }
+}
